% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_network.R
\name{grab_network}
\alias{grab_network}
\title{Grab a network of coauthors from Google Scholar profile}
\usage{
grab_network(scholar_id, n_deep = 1)
}
\arguments{
\item{scholar_id}{The endline of a google scholar profile. For example, 'citations?user=wKNoabEAAAAJ&hl=en&oi=ao' comes
from the profile https://scholar.google.com/citations?user=wKNoabEAAAAJ&hl=en&oi=ao}

\item{n_deep}{The number of degrees that you want to go down the network. When \code{n_deep} is equal to \code{1}
then \code{grab_network} will only grab the coauthors of Joe and Mary, so Joe -- > Mary --> All coauthors. This can get
out of control very quickly as the network will grow exponentially. I suggest the reader to not use a number more
than \code{2} and the preferred number would be \code{1}, the default.}
}
\value{
A \code{tibble} with the all authors and coauthors.
}
\description{
Grab a network of coauthors from Google Scholar profile
}
\examples{

\dontrun{

final_network <- grab_network('citations?user=wKNoabEAAAAJ&hl=en&oi=ao')
plot_coauthors(final_network)

}

}
