% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coauthors.R
\name{plot_coauthors}
\alias{plot_coauthors}
\title{Plot a network of coauthors}
\usage{
plot_coauthors(network, size_labels = 5)
}
\arguments{
\item{network}{A data frame given by \code{grab_network}}

\item{size_labels}{Size of the label names}
}
\value{
a \code{ggplot2} object but prints a plot as a side effect.
}
\description{
Plot a network of coauthors
}
\examples{

\dontrun{
final_network <- grab_network('citations?user=amYIKXQAAAAJ&hl=en')
plot_coauthors(final_network)
}
}
